[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,7 +2,7 @@
 using namespace std;
 
 void A(int a, int b){
-    cout << (a+b)*(a-b);
+    cout << (long)(a+b)*(a-b);
 }
 
 int main(){
